Fix dung306080_controlnet: force dtype cast after from_single_file to fix BFloat16/Float32 mismatch

+++ dung306080_controlnet/pytorch/loader.py
@@ -135,7 +135,7 @@ class ModelLoader(ForgeModel):
             model_path,
             config=_CONFIGS[version],
             torch_dtype=dtype,
-        )
+        ).to(dtype=dtype)
         self._controlnet.eval()
         return self._controlnet
